Added transform for position

+++ include/PoseFrame.h
@@ -2,12 +2,12 @@ class PoseFrame {
 
 public:
 
-    float x;
-    float y;
-    float orientation;
+    double x;
+    double y;
+    double orientation;
 
     PoseFrame() : x(0), y(0), orientation(0) { }
-    PoseFrame(float x, float y, float orientation) 
+    PoseFrame(double x, double y, double orientation) 
     {
         this->x = x; 
         this->y = y;
